surogaAPI/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
certifi==2022.6.15.2
cffi==1.17.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
Deprecated==1.2.14
Django==4.1.7
django-decouple==2.1
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.4
importlib_metadata==8.4.0
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.4.1
MarkupSafe==3.0.1
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.1.0
microsoft-kiota-http==1.3.4
microsoft-kiota-serialization-form==0.1.1
microsoft-kiota-serialization-json==1.3.3
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.1.5
msgraph-sdk==1.9.0
mssql-django==1.5
multidict==6.1.0
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
pendulum==3.0.0
portalocker==2.10.1
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.3
std-uritemplate==2.0.0
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.2.0
wrapt==1.16.0
yarl==1.15.1
zipp==1.0.0
